Refactoring for Multiple Client Fix
Changes:
- EntityData will supply a integer id of the player in respect to the id supplied in the lobby. Ensuring correct data for each player entity
- Health and time_damaged added to EntityData
- Disabled multiple lobby feature until the implementation of LobbyData updates
- UDP port of lobby inputs no longer an ephemeral port.

+++ src/main/java/configs/StatsConfig.java
@@ -1,7 +1,7 @@
 package configs;
 
 public class StatsConfig {
-    public static final double PLAYER_HEALTH = 100;
+    public static final int PLAYER_HEALTH = 100;
     public static final double PROJECTILE_DAMAGE = 10;
 
     // Speed per millisecond
